requirements.txt
fastapi
uvicorn
python-multipart
openpyxl
langchain-community
langchain-gigachat
python-dotenv
pandas
